Fix three 64-bit portability issues found in warning analysis
- Fix UB in buildObjWithLoadedSlots: change parameter type from
  long slotv[100][2] to long (*slotv)[100][2] to match the caller
  pointer type, eliminating incompatible pointer types warning
- Widen Folder.nItems and Folder.allocedItems from short to int
  to prevent overflow in realloc size calculations
- Add explicit (int) casts for intptr_t-to-int conversions in
  getClassInfoByID/getClassInfoByName call sites

+++ src/vw/catalog.h
@@ -72,8 +72,8 @@ typedef struct folderStruct {
 
     char* commentURL;
     Item** items;
-    short nItems;
-    short allocedItems;
+    int nItems;
+    int allocedItems;
     short cw, ch; /* size of folder canvas area (when open) */
 
     XmString nItemsXMS;
